+++ README.md
@@ -9,32 +9,32 @@ Annotation based AWS Lambda extension implementation for Ballerina.
 
 ## Supported Annotations:
 
-### @awslambda:Function
+### @lambda:Function
 - Supported with Ballerina package level functions
 
 ### Annotation Usage Sample:
 
 ```ballerina
-import ballerinax/awslambda;
+import ballerinax/aws.lambda;
 import ballerina/uuid;
 import ballerina/io;
 
-// The `@awslambda:Function` annotation marks a function to
+// The `@lambda:Function` annotation marks a function to
 // generate an AWS Lambda function
-@awslambda:Function
-public function echo(awslambda:Context ctx, json input) returns json {
+@lambda:Function
+public function echo(lambda:Context ctx, json input) returns json {
    return input;
 }
 
-@awslambda:Function
-public function uuid(awslambda:Context ctx, json input) returns json {
+@lambda:Function
+public function uuid(lambda:Context ctx, json input) returns json {
    return uuid:createType1AsString();
 }
 
-// The `awslambda:Context` object contains request execution
+// The `lambda:Context` object contains request execution
 // context information
-@awslambda:Function
-public function ctxinfo(awslambda:Context ctx, json input) returns json|error {
+@lambda:Function
+public function ctxinfo(lambda:Context ctx, json input) returns json|error {
    json result = { RequestID: ctx.getRequestId(),
                    DeadlineMS: ctx.getDeadlineMs(),
                    InvokedFunctionArn: ctx.getInvokedFunctionArn(),
@@ -43,33 +43,33 @@ public function ctxinfo(awslambda:Context ctx, json input) returns json|error {
    return result;
 }
 
-@awslambda:Function
-public function notifySQS(awslambda:Context ctx, 
-                          awslambda:SQSEvent event) returns json {
+@lambda:Function
+public function notifySQS(lambda:Context ctx, 
+                          lambda:SQSEvent event) returns json {
     return event.Records[0].body;
 }
 
-@awslambda:Function
-public function notifyS3(awslambda:Context ctx, 
-                         awslambda:S3Event event) returns json {
+@lambda:Function
+public function notifyS3(lambda:Context ctx, 
+                         lambda:S3Event event) returns json {
     return event.Records[0].s3.'object.key;
 }
 
-@awslambda:Function
-public function notifyDynamoDB(awslambda:Context ctx, 
-                               awslambda:DynamoDBEvent event) returns json {
+@lambda:Function
+public function notifyDynamoDB(lambda:Context ctx, 
+                               lambda:DynamoDBEvent event) returns json {
     return event.Records[0].dynamodb.Keys.toString();
 }
 
-@awslambda:Function
-public function notifySES(awslambda:Context ctx, 
-                          awslambda:SESEvent event) returns json {
+@lambda:Function
+public function notifySES(lambda:Context ctx, 
+                          lambda:SESEvent event) returns json {
     return event.Records[0].ses.mail.commonHeaders.subject;
 }
 
-@awslambda:Function
-public function apigwRequest(awslambda:Context ctx, 
-                             awslambda:APIGatewayProxyRequest request) {
+@lambda:Function
+public function apigwRequest(lambda:Context ctx, 
+                             lambda:APIGatewayProxyRequest request) {
     io:println("Path: ", request.path);
 }
 ```
@@ -83,7 +83,7 @@ Compiling source
 
 Generating executables
 	functions.jar
-	@awslambda:Function: echo, uuid, ctxinfo, notifySQS, notifyS3, notifyDynamoDB, notifySES, apigwRequest
+	@aws.lambda:Function: echo, uuid, ctxinfo, notifySQS, notifyS3, notifyDynamoDB, notifySES, apigwRequest
 
 	Run the following command to deploy each Ballerina AWS Lambda function:
 	aws lambda create-function --function-name <FUNCTION_NAME> --zip-file fileb://aws-ballerina-lambda-functions.zip --handler functions.<FUNCTION_NAME> --runtime provided --role <LAMBDA_ROLE_ARN> --layers arn:aws:lambda:<REGION_ID>:134633749276:layer:ballerina-jre11:6
