+++ build-config/resources/CompilerPlugin.toml
@@ -1,6 +1,6 @@
 [plugin]
 id = "awslambda"
-class = "org.ballerinax.awslambda.generator.AWSLambdaCompilerPlugin"
+class = "org.ballerinax.aws.lambda.generator.AWSLambdaCompilerPlugin"
 
 [[dependency]]
 path = "../compiler-plugin/build/libs/aws.lambda-compiler-plugin-@project.version@.jar"
