--- compiler-plugin-tests/src/test/java/org/ballerinax/awslambda/generator/test/utils/ParserTestConstants.java
@@ -1,42 +1,0 @@
-/*
- * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.ballerinax.awslambda.generator.test.utils;
-
-/**
- * Constants related to the parser.
- * 
- * @since 1.2.0
- */
-public class ParserTestConstants {
-
-    public static final String KIND_FIELD = "kind";
-
-    public static final String CHILDREN_FIELD = "children";
-    public static final String DIAGNOSTICS_FIELD = "diagnostics";
-
-    public static final String VALUE_FIELD = "value";
-    public static final String INVALID_NODE_FIELD = "invalidNode";
-
-    public static final String IS_MISSING_FIELD = "isMissing";
-    public static final String HAS_DIAGNOSTICS = "hasDiagnostics";
-
-    public static final String LEADING_MINUTIAE = "leadingMinutiae";
-
-    public static final String TRAILING_MINUTIAE = "trailingMinutiae";
-
-}
