+++ compiler-plugin/src/main/java/org/ballerinax/aws/lambda/generator/LambdaHandlerContainer.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.ballerinax.aws.lambda.generator;
+
+import io.ballerina.compiler.syntax.tree.FunctionDefinitionNode;
+
+import java.util.List;
+
+/**
+ * Represents a container for lambda functions in a module.
+ *
+ * @since 2.0.0
+ */
+public class LambdaHandlerContainer {
+    private final List<FunctionDefinitionNode> functions;
+
+    public LambdaHandlerContainer(List<FunctionDefinitionNode> functions) {
+        this.functions = functions;
+    }
+
+    public List<FunctionDefinitionNode> getFunctions() {
+        return functions;
+    }
+}
