--- compiler-plugin/src/main/java/org/ballerinax/awslambda/generator/LambdaLifecycleListener.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.ballerinax.awslambda.generator;
-
-import io.ballerina.projects.plugins.CompilerLifecycleContext;
-import io.ballerina.projects.plugins.CompilerLifecycleListener;
-import org.ballerinax.awslambda.generator.tasks.LambdaCodeGeneratedTask;
-
-/**
- * Contains the code generation part of the AWS Lambda.
- *
- * @since 2.0.0
- */
-public class LambdaLifecycleListener extends CompilerLifecycleListener {
-    @Override
-    public void init(CompilerLifecycleContext compilerLifecycleContext) {
-        compilerLifecycleContext.addCodeGenerationCompletedTask(new LambdaCodeGeneratedTask());
-    }
-}
